mrhistmatch: Fix compilation on Mac

+++ cmd/mrhistmatch.cpp
@@ -76,24 +76,24 @@ void match_linear (Image<float>& input,
     ProgressBar progress ("Loading & sorting image data", 4);
     if (mask_input.valid()) {
       for (auto l = Loop(input) (input, mask_input); l; ++l) {
-        if (mask_input.value() && std::isfinite (input.value()))
+        if (mask_input.value() && std::isfinite (static_cast<float>(input.value())))
           input_data.push_back (input.value());
       }
     } else {
       for (auto l = Loop(input) (input); l; ++l) {
-        if (std::isfinite (input.value()))
+        if (std::isfinite (static_cast<float>(input.value())))
           input_data.push_back (input.value());
       }
     }
     ++progress;
     if (mask_target.valid()) {
       for (auto l = Loop(target) (target, mask_target); l; ++l) {
-        if (mask_target.value() && std::isfinite (target.value()))
+        if (mask_target.value() && std::isfinite (static_cast<float>(target.value())))
           target_data.push_back (target.value());
       }
     } else {
       for (auto l = Loop(target) (target); l; ++l) {
-        if (std::isfinite (target.value()))
+        if (std::isfinite (static_cast<float>(target.value())))
           target_data.push_back (target.value());
       }
     }
@@ -178,7 +178,7 @@ void match_nonlinear (Image<float>& input,
   H.datatype().set_byte_order_native();
   auto output = Image<float>::create (argument[3], H);
   for (auto l = Loop("Writing output data", input) (input, output); l; ++l) {
-    if (std::isfinite(static_cast<float>(input.value()))) {
+    if (std::isfinite (static_cast<float>(input.value()))) {
       output.value() = matcher (input.value());
     } else {
       output.value() = input.value();
